`soliloquy` renamed to `kartoshka`
...in case you couldn't tell I've been really enjoying ep 1 of ZATO lol... and also soviet choco potato is so intriguing !

I really love the simple yet evocative storytelling structure of the specific style of renpy that's used, and hoping i can either salvage-veer this if the 3D aspect flounders into the easier and text-driven 2D realm (want to experiment with this style!!) or incorporate both for a unique storytelling tool; general feel is to hopefully use some of that lisp metamagic to make branching and merging narrative easier, hot-swap shenanigans, etc etc. A nice dip into lisp land!